first/requirements.txt
backports.csv==1.0.7
configparser==3.7.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-filter==2.2.0
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
pyPdf==1.13
pytz==2019.1
PyYAML==5.1
sorter==0.5
sqlparse==0.3.0
tablib==0.13.0
webstack-django-sorting==1.0.1
xlrd==1.2.0
xlwt==1.3.0
